requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# MCP Protocol
mcp>=1.0.0

# HTTP Client
httpx>=0.27.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
ruff==0.1.6
